Reorganize into spec
