+++ sealedenumprocessor/src/main/kotlin/com/livefront/sealedenum/internal/spec/SealedEnumNamePropertySpec.kt
@@ -1,4 +1,4 @@
-package com.livefront.sealedenum.internal
+package com.livefront.sealedenum.internal.spec
 
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FunSpec
@@ -6,22 +6,21 @@ import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeName
 import javax.lang.model.element.TypeElement
 
-internal data class SealedEnumEnumPropertySpec(
+internal data class SealedEnumNamePropertySpec(
     private val sealedClass: SealedClass,
     private val parameterizedSealedClass: TypeName,
     private val sealedClassCompanionObjectElement: TypeElement,
     private val sealedEnum: ClassName,
-    private val enumForSealedEnum: ClassName,
     private val enumPrefix: String
 ) {
     fun build(): PropertySpec {
-        val propertySpecBuilder = PropertySpec.builder(pascalCaseToCamelCase(enumPrefix + "Enum"), enumForSealedEnum)
+        val propertySpecBuilder = PropertySpec.builder(pascalCaseToCamelCase(enumPrefix + "Name"), String::class)
             .addOriginatingElement(sealedClassCompanionObjectElement)
-            .addKdoc("The isomorphic [%T] for [this].", enumForSealedEnum)
+            .addKdoc("The name of [this] for use with valueOf.")
             .receiver(parameterizedSealedClass)
             .getter(
                 FunSpec.getterBuilder()
-                    .addStatement("return %T.sealedObjectToEnum(this)", sealedEnum)
+                    .addStatement("return %T.nameOf(this)", sealedEnum)
                     .build()
             )
