+++ sealedenumprocessor/src/main/kotlin/com/livefront/sealedenum/internal/spec/SealedEnumOrdinalPropertySpec.kt
@@ -1,4 +1,4 @@
-package com.livefront.sealedenum.internal
+package com.livefront.sealedenum.internal.spec
 
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.FunSpec
@@ -6,21 +6,21 @@ import com.squareup.kotlinpoet.PropertySpec
 import com.squareup.kotlinpoet.TypeName
 import javax.lang.model.element.TypeElement
 
-internal data class EnumSealedObjectPropertySpec(
+internal data class SealedEnumOrdinalPropertySpec(
     private val sealedClass: SealedClass,
     private val parameterizedSealedClass: TypeName,
     private val sealedClassCompanionObjectElement: TypeElement,
     private val sealedEnum: ClassName,
-    private val enumForSealedEnum: ClassName
+    private val enumPrefix: String
 ) {
     fun build(): PropertySpec {
-        val propertySpecBuilder = PropertySpec.builder("sealedObject", parameterizedSealedClass)
+        val propertySpecBuilder = PropertySpec.builder(pascalCaseToCamelCase(enumPrefix + "Ordinal"), Int::class)
             .addOriginatingElement(sealedClassCompanionObjectElement)
-            .addKdoc("The isomorphic [%T] for [this].", sealedClass)
-            .receiver(enumForSealedEnum)
+            .addKdoc("The index of [this] in the values list.")
+            .receiver(parameterizedSealedClass)
             .getter(
                 FunSpec.getterBuilder()
-                    .addStatement("return %T.enumToSealedObject(this)", sealedEnum)
+                    .addStatement("return %T.ordinalOf(this)", sealedEnum)
                     .build()
             )
