+++ sealedenumprocessor/src/main/kotlin/com/livefront/sealedenum/internal/spec/SealedEnumTypeSpec.kt
@@ -1,4 +1,4 @@
-package com.livefront.sealedenum.internal
+package com.livefront.sealedenum.internal.spec
 
 import com.livefront.sealedenum.EnumForSealedEnumProvider
 import com.livefront.sealedenum.SealedEnum
